requirements.txt
# LiveKit core SDK
livekit-api

# Agents and plugins
livekit-agents
livekit-plugins-groq
livekit-plugins-cartesia
livekit-plugins-noise-cancellation
livekit-plugins-silero
livekit-plugins-turn-detector
huggingface_hub

# Flask + async support
flask[async]
flask-cors

# Env and utils
python-dotenv
requests
uuid

# AI / ML
pydantic-ai-slim[openai]

# Timezone
tzdata
